refactor: 주요관심사 api 훅 분리
